+++ R/proxy.R
@@ -22,11 +22,13 @@
 #' @param diag if \code{TRUE}, only compute diagonal elements of the
 #'   similarity/distance matrix; useful when comparing corresponding rows or
 #'   columns of `x` and `y`.
-#' @param use_nan if \code{TRUE}, return `NaN` if the standard deviation of a
+#' @param use_nan if `TRUE`, return `NaN` if the standard deviation of a
 #'   vector is zero when `method` is "correlation"; if all the values are zero
-#'   in a vector when `method` is "cosine", "kullback", "jeffreys" or "chisquared".
-#'   Note that use of `NaN` makes the similarity/distance matrix denser and
-#'   therefore larger.
+#'   in a vector when `method` is "cosine", "chisquared", "kullback", "jeffreys"
+#'   or "jensen". Note that use of `NaN` makes the similarity/distance matrix
+#'   denser and therefore larger in RAM. If `FALSE`, return zero in same use
+#'   situations as above. If `NULL`, will also return zero but also generate
+#'   a warning (default).
 #' @param digits determines rounding of small values towards zero. Use primarily
 #'   to correct rounding errors in C++. See \link{zapsmall}.
 #' @details
@@ -67,7 +69,7 @@ simil <- function(x, y = NULL, margin = 1,
                   method = c("cosine", "correlation", "jaccard", "ejaccard",
                              "dice", "edice", "hamann", "faith", "simple matching"),
                   min_simil = NULL, rank = NULL, drop0 = FALSE, diag = FALSE,
-                  use_nan = FALSE, digits = 14) {
+                  use_nan = NULL, digits = 14) {
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
@@ -87,7 +89,7 @@ simil <- function(x, y = NULL, margin = 1,
 dist <- function(x, y = NULL, margin = 1,
                  method = c("euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                             "manhattan", "maximum", "canberra", "minkowski", "hamming"),
-                 p = 2, smooth = 0, drop0 = FALSE, diag = FALSE, use_nan = FALSE, digits = 14) {
+                 p = 2, smooth = 0, drop0 = FALSE, diag = FALSE, use_nan = NULL, digits = 14) {
 
     method <- match.arg(method)
     proxy(x, y, margin, method, p = p, smooth = smooth, drop0 = drop0,
@@ -102,18 +104,19 @@ proxy <- function(x, y = NULL, margin = 1,
                              "euclidean", "chisquared", "kullback", "jeffreys", "jensen",
                              "manhattan", "maximum", "canberra", "minkowski", "hamming"),
                   p = 2, smooth = 0, min_proxy = NULL, rank = NULL, drop0 = FALSE,
-                  diag = FALSE, use_nan = FALSE, digits = 14) {
+                  diag = FALSE, use_nan = NULL, digits = 14) {
 
     method[method == "hamman"] <- "hamann" # for transition
     method <- match.arg(method)
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
-
+    #x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(as(x, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     symm <- is.null(y)
 
     if (is.null(y)) {
         y <- x
     } else {
-        y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        #y <- as(as(y, "CsparseMatrix"), "dgCMatrix")
+        y <- as(as(as(y, "CsparseMatrix"), "generalMatrix"), "dMatrix") # for Matrix v1.4-2 or later
     }
     if (!margin %in% c(1, 2))
         stop("Matrgin must be 1 (row) or 2 (column)")
@@ -132,14 +135,21 @@ proxy <- function(x, y = NULL, margin = 1,
         rank <- ncol(x)
     if (rank < 1)
         stop("rank must be great than or equal to 1")
-    if (!use_nan) {
-        if (method == "correlation") {
-            if (any(colSds(x) == 0) || any(colSds(y) == 0))
-                warning("x or y has vectors with zero standard deviation; consider setting use_nan = TRUE", call. = FALSE)
-        } else if (method %in% c("cosine", "kullback", "chisquared", "jeffreys", "jensen")) {
-            if (any(colZeros(x) == nrow(x)) || any(colZeros(y) == nrow(y)))
-                warning("x or y has vectors with all zero; consider setting use_nan = TRUE", call. = FALSE)
+    if (is.null(use_nan)) {
+        if (method == "correlation" && (any(colSds(x) == 0) || any(colSds(y) == 0))) {
+            warning(paste0(
+                "x or y has vectors with zero standard deviation; ",
+                "consider setting use_nan = TRUE to set these values to NaN ",
+                "or use_nan = FALSE to suppress this warning"), call. = FALSE)
+        } else if (
+            method %in% c("cosine", "kullback", "chisquared", "jeffreys", "jensen") &&
+            (any(colZeros(x) == nrow(x)) || any(colZeros(y) == nrow(y)))) {
+            warning(paste0(
+                "x or y has vectors with all zero; ",
+                "consider setting use_nan = TRUE to set these values to NaN ",
+                "or use_nan = FALSE to suppress this warning"), call. = FALSE)
         }
+        use_nan <- FALSE
     }
     boolean <- FALSE
     weight <- 1
@@ -165,8 +175,8 @@ proxy <- function(x, y = NULL, margin = 1,
         weight <- p
     }
     if (boolean) {
-        x <- as(as(x, "lgCMatrix"), "dgCMatrix")
-        y <- as(as(y, "lgCMatrix"), "dgCMatrix")
+        x <- as(as(x, "lMatrix"), "dMatrix")
+        y <- as(as(y, "lMatrix"), "dMatrix")
     }
     if (method %in% c("cosine", "correlation", "euclidean") && !diag) {
         result <- cpp_linear(
@@ -217,7 +227,7 @@ proxy <- function(x, y = NULL, margin = 1,
 #' apply(mt, 2, sd) # the same
 #' @export
 colSds <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- cpp_sd(x)
     names(result) <- colnames(x)
     return(result)
@@ -226,7 +236,7 @@ colSds <- function(x) {
 #' @rdname colSds
 #' @export
 rowSds <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- cpp_sd(t(x))
     names(result) <- rownames(x)
     return(result)
@@ -242,7 +252,7 @@ rowSds <- function(x) {
 #' apply(mt, 2, function(x) sum(x == 0)) # the same
 #' @export
 colZeros <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- nrow(x) - cpp_nz(x)
     names(result) <- colnames(x)
     return(result)
@@ -251,7 +261,7 @@ colZeros <- function(x) {
 #' @rdname colZeros
 #' @export
 rowZeros <- function(x) {
-    x <- as(as(x, "CsparseMatrix"), "dgCMatrix")
+    x <- as(as(x, "CsparseMatrix"), "dMatrix")
     result <- ncol(x) - cpp_nz(t(x))
     names(result) <- rownames(x)
     return(result)